Rename validation namespace
