+++ packages/framework/src/Framework/Features/Publications/PublicationFieldTypes.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Hyde\Framework\Features\Publications;
 
-use Hyde\Framework\Features\Publications\Rules\BooleanRule;
+use Hyde\Framework\Features\Publications\Validation\BooleanRule;
 use Illuminate\Support\Collection;
 
 /**
