+++ packages/framework/src/Framework/Features/Publications/Validation/BooleanRule.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Hyde\Framework\Features\Publications\Rules;
+namespace Hyde\Framework\Features\Publications\Validation;
 
 use Illuminate\Contracts\Validation\InvokableRule;
 use function in_array;
